Return meta data for matches API
In order to properly allow pagination on the client-side, the matches
api now returns some metadata. This includes the total amount of matches
matching the filter criteria, as well as on which page the query is
right now.

This is a breaking change in the api, as the response structure is
changed to support the additional information.

Fixes #95