fix: wrong calculations for coop matches
Bug description: in coop calculations the a factor has been computed and passed as number of matches ... this a massive bug, because teams with less than 30 matches only get the a=20 multiplier in the final calculations, because 40 has been passed as number of matches. Until now, all coop matches with clans that have less than 30 matches only got half the points they descerved. This is irreversible unfortunately.